Add libak.h header and fix includes in libak.c
- Create src/agentic/libak.h with complete C API definitions for userspace Authority Kernel access
- Includes syscall numbers, error codes, data structures, and function declarations
- Fix libak.c include order to properly reference libak.h
- Revert heap.h pointer casting back to u64 (not uintptr_t) for Nanos boot code compatibility
- Enables builds to pass on macOS with both kernel and libak binaries

+++ src/runtime/heap/heap.h
@@ -66,9 +66,8 @@ backed_heap reserve_backed_heap_wrapper(heap meta, backed_heap parent, bytes res
 
 static inline void *page_of(void *x, bytes pagesize)
 {
-    /* Use uintptr_t for portable pointer-to-integer conversions */
-    return((void *)((uintptr_t)x &
-                    (~((uintptr_t)pagesize-1))));
+    return((void *)((u64)x &
+                    (~((u64)pagesize-1))));
 }
 
 static inline int subdivide(int quantum, int per, int s, int o)
